+++ v2/googlecloud-to-snowflake/pom.xml
@@ -22,79 +22,16 @@
 	</parent>
 
 	<dependencies>
-
-
-		<dependency>
-			<groupId>org.apache.beam</groupId>
-			<artifactId>beam-sdks-java-io-snowflake</artifactId>
-			<version>${beam.version}</version>
-		</dependency>
-
-
-		<dependency>
-			<groupId>com.fasterxml.jackson.core</groupId>
-			<artifactId>jackson-databind</artifactId>
-			<version>2.9.4</version>
-		</dependency>
-
-		<dependency>
-			<groupId>com.google.http-client</groupId>
-			<artifactId>google-http-client</artifactId>
-			<version>1.39.2</version>
-		</dependency>
-		<dependency>
-			<groupId>com.google.api-client</groupId>
-			<artifactId>google-api-client</artifactId>
-			<version>1.32.1</version>
-		</dependency>
-		<dependency>
-			<groupId>com.google.guava</groupId>
-			<artifactId>guava</artifactId>
-			<version>29.0-jre</version>
-		</dependency>
-
-		<dependency>
-			<groupId>io.grpc</groupId>
-			<artifactId>grpc-netty</artifactId>
-			<version>1.32.2</version>
-		</dependency>
-
-		<!-- google-cloud-kms -->
-		<dependency>
-			<groupId>com.google.cloud</groupId>
-			<artifactId>google-cloud-kms</artifactId>
-			<version>1.40.0</version>
-			<exclusions>
-				<exclusion>
-					<groupId>com.google.guava</groupId>
-					<artifactId>guava</artifactId>
-				</exclusion>
-			</exclusions>
-		</dependency>
-
-
 		<dependency>
 			<groupId>com.google.cloud.teleport.v2</groupId>
 			<artifactId>Snowflake-common</artifactId>
 			<version>1.0-SNAPSHOT</version>
 			<scope>compile</scope>
-
 		</dependency>
-		<dependency>
-			<groupId>com.google.cloud.teleport.v2</groupId>
-			<artifactId>common</artifactId>
-			<version>1.0-SNAPSHOT</version>
-			<scope>compile</scope>
-
-		</dependency>
-
 	</dependencies>
 
 	<build>
-
 		<directory>${mvn-target-dir}</directory>
-
 	</build>
-
 
 </project>
