--- v2/googlecloud-to-snowflake/src/main/resources/PubSubToSnowflake_metadata.json
@@ -1,101 +1,0 @@
-{
-  "name": "PubSubSubscriptiontoSnowflake",
-  "parameters": [
-    {
-     "name": "serverName",
-      "helpText": "Snowflake Server Name",
-      "label": "Enter Snowflake Server Name "
-    },
-    {
-      "name": "tokenKMSEncryptionKey",
-      "helpText": "KMS Encryption Key",
-      "label": "Enter KMS Encryption Key (format: projects/{gcp_project}/locations/{key_region}/keyRings/{key_ring}/cryptoKeys/{kms_key_name})"
-    },
-    {
-      "name": "username",
-      "helpText": "Snowflake User Name",
-      "label": "Enter Snowflake User Name "
-    },
-    {
-      "name": "rawPrivateKey",
-      "helpText": "Raw private key",
-      "label": "Enter raw private key string"
-    },
-    {
-      "name": "privateKeyPassphrase",
-      "helpText": "Passphrase for the private key",
-      "label": "Enter Passphrase for the private key"
-    },
-    {
-      "name": "storageIntegrationName",
-      "helpText": "Snowflake's storage integration name",
-      "label": "Enter Snowflake's storage integration name "
-    },
-    {
-      "regexes": [
-        "^[^\\n\\r]+$"
-      ],
-      "name": "snowPipe",
-      "helpText": "Snowflake's snowPipe name",
-      "label": "Snowflake's snowPipe name"
-    },
-    {
-      "name": "role",
-      "helpText": "Snowflake role",
-      "label": "Enter Snowflake role "
-    },
-    {
-      "name": "warehouse",
-      "helpText": "Snowflake's warehouse name",
-      "label": "Enter Snowflake's warehouse name "
-    },
-    {
-      "name": "database",
-      "helpText": "Snowflake's database name",
-      "label": "Enter Snowflake's database name "
-    },
-    {
-      "name": "schema",
-      "helpText": "Snowflake's database schema name",
-      "label": "Enter Snowflake's database schema name "
-    },
-    {
-      "name": "table",
-      "helpText": "Snowflake's table name",
-      "label": "Enter Snowflake's table name"
-    },
-    {
-      "name": "inputSubscription",
-      "helpText": "Name of the Pub/Sub Subscription",
-      "label": "Enter name of the pub/sub subscription"
-    },
-    {
-      "name": "stagingBucketName",
-      "helpText": "Staging bucket location starting with (gs://) for data processing micro-batches",
-      "label": "Enter staging bucket location starting with (gs://) for data processing micro-batches"
-    },
-	{
-      "isOptional": true,
-      "name": "oauthToken",
-      "helpText": "OAuth token",
-      "label": "Enter OAuth token to set OAuth authentication."
-    },
-	{
-      "isOptional": true,
-      "name": "password",
-      "helpText": "Password",
-      "label": "Enter Password to set username/password authentication."
-    },
-    {
-      "isOptional": true,
-      "name": "gcpTempLocation",
-      "helpText": "Temporary location in GCP",
-      "label": "Enter Temporary location in GCP."
-    },
-    {
-      "name": "sourceFormat",
-      "helpText": "Source format of pubsub data",
-      "label": "Enter source format of pubsub data."
-    }
-  ]
-}
