+++ v2/googlecloud-to-snowflake/src/main/resources/googlecloud-to-snowflake-command-spec.json
