--- v2/snowflake-common/Readme.Snowflake-Common.md
@@ -1,19 +1,0 @@
-# Snowflake Common
-
-The Snowflake common utilities are built for the use of dataflow pipelines to transfer data from GoogleCloud to Snowflake and vice versa.
-
-**Requirements**
-
-* Java 8
-* Maven
-
-**Getting Started**
-
-These common utilites are currently used by the Dataflow template for Snowflake under project GoogleCloud-to-Snowflake (DataflowTemplates/v2/GoogleCloud-to-Snowflake). 
-
-**Included Functionality**
-
-The included utilities provides following functionality:
-
-* Decrypt the encrypted value for parameters
-* Data Mapper for source formats (CSV, JSON)
