+++ v2/snowflake-common/pom.xml
@@ -21,38 +21,22 @@
 	</parent>
 
 	<dependencies>
-
-
 		<dependency>
 			<groupId>org.apache.beam</groupId>
 			<artifactId>beam-sdks-java-io-snowflake</artifactId>
 			<version>${beam.version}</version>
 		</dependency>
-		<dependency>
-			<groupId>io.grpc</groupId>
-			<artifactId>grpc-netty</artifactId>
-			<version>1.32.2</version>
-		</dependency>
 
-		<!-- google-cloud-kms -->
 		<dependency>
-			<groupId>com.google.cloud</groupId>
-			<artifactId>google-cloud-kms</artifactId>
-			<version>1.40.0</version>
-			<exclusions>
-				<exclusion>
-					<groupId>com.google.guava</groupId>
-					<artifactId>guava</artifactId>
-				</exclusion>
-			</exclusions>
+			<groupId>com.google.cloud.teleport.v2</groupId>
+			<artifactId>common</artifactId>
+			<version>1.0-SNAPSHOT</version>
+			<scope>compile</scope>
 		</dependency>
-
 	</dependencies>
 
 	<build>
 		<directory>${mvn-target-dir}</directory>
-
 	</build>
-
 
 </project>
