+++ v2/snowflake-common/src/main/java/com/google/cloud/teleport/v2/snowflake/util/DecryptUtil.java
@@ -17,12 +17,14 @@ package com.google.cloud.teleport.v2.snowflake.util;
 
 import com.google.cloud.kms.v1.DecryptResponse;
 import com.google.cloud.kms.v1.KeyManagementServiceClient;
+import com.google.common.base.Strings;
 import com.google.protobuf.ByteString;
 import java.io.IOException;
 import java.util.Base64;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import org.apache.beam.sdk.options.ValueProvider;
+import org.apache.beam.sdk.options.ValueProvider.StaticValueProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -33,9 +35,9 @@ import org.slf4j.LoggerFactory;
  */
 public class DecryptUtil {
 
-	private static final Pattern KEYNAME_PATTERN = Pattern
-			.compile("projects/([^/]+)/locations/([a-zA-Z0-9_-]{1,63})/keyRings/"
-					+ "[a-zA-Z0-9_-]{1,63}/cryptoKeys/[a-zA-Z0-9_-]{1,63}");
+	private static final Pattern KEY_NAME_PATTERN = Pattern.compile("projects/([^/]+)/locations/([a-zA-Z0-9_-]{1,63})/keyRings/[a-zA-Z0-9_-]{1,63}/cryptoKeys/[a-zA-Z0-9_-]{1,63}");
+
+	private static final Pattern ENCRYPT_PATTERN = Pattern.compile("\\$\\{(.*?)}");
 
 	/** The log to output status messages to. */
 	private static final Logger LOG = LoggerFactory.getLogger(DecryptUtil.class);
@@ -50,37 +52,32 @@ public class DecryptUtil {
 	 * this KMS API Encrypt Link</a>.
 	 * 
 	 * @param encryptedValue base64 encrypted string
-	 * @param kmsKey encryption key 
+	 * @param kmsKey encryption key
 	 * @return decrypted value
 	 * @throws IOException
 	 */
 
-	public static String decryptWithKMS(String encryptedValue, String kmsKey) throws IOException {
-		/*
-		 * kmsKey should be in the following format:
-		 * projects/{gcp_project}/locations/{key_region}/keyRings/{key_ring}/cryptoKeys/
-		 * {kms_key_name}
-		 */
+	private String kmsKey;
+	private KeyManagementServiceClient keyManagementServiceClient;
 
-		if (kmsKey == null || encryptedValue.isEmpty()) {
-			LOG.info("KMS Key is not specified. Using: " + encryptedValue);
-			return encryptedValue;
-		} else if (!validateKmsKey(kmsKey)) {
+	public DecryptUtil(String kmsKey) throws IOException {
+		if (Strings.isNullOrEmpty(kmsKey)|| !validateKmsKey(kmsKey)){
 			IllegalArgumentException exception = new IllegalArgumentException("Provided KMS Key %s is invalid");
 			throw new RuntimeException(exception);
-		} else {
-
-			byte[] cipherText = Base64.getDecoder().decode(encryptedValue.getBytes("UTF-8"));
-
-			try (KeyManagementServiceClient client = KeyManagementServiceClient.create()) {
-
-				// Decrypt the ciphertext with Cloud KMS.
-				DecryptResponse response = client.decrypt(kmsKey, ByteString.copyFrom(cipherText));
-
-				// Extract the plaintext from the response.
-				return new String(response.getPlaintext().toByteArray());
-			}
 		}
+		this.kmsKey = kmsKey;
+		this.keyManagementServiceClient = KeyManagementServiceClient.create();
+	}
+
+	private String decryptWithKMS(String encryptedValue) throws IOException {
+
+		byte[] cipherText = Base64.getDecoder().decode(encryptedValue.getBytes("UTF-8"));
+
+		// Decrypt the ciphertext with Cloud KMS.
+		DecryptResponse decryptResponse = this.keyManagementServiceClient.decrypt(this.kmsKey, ByteString.copyFrom(cipherText));
+
+		// Extract the plaintext from the response.
+		return new String(decryptResponse.getPlaintext().toByteArray());
 	}
 
 	/**
@@ -92,118 +89,71 @@ public class DecryptUtil {
 	 * @return true if key matches the pattern above else returns false
 	 */
 	private static boolean validateKmsKey(String kmsKey) {
-		return KEYNAME_PATTERN.matcher(kmsKey).matches();
+		return KEY_NAME_PATTERN.matcher(kmsKey).matches();
 	}
 
 	/**
-	 * Provides the raw value which is wrapped by the ValueProvider if the value is accessible. If the value is 
-	 * not accessible at execution time then empty string is returned("").
+	 * Checks if pipeline parameter is encrypted. If pipeline parameter is
+	 * encrypted, returns the decrypted value using kms encryption key otherwise
+	 * returns the raw value as is.
 	 * 
-	 * @param <T> any subType of ValueProvider interface
-	 * @param t a ValueProvider type which wraps the raw value provided at execution phase
-	 * @return raw value wrapped by the ValueProvider type.
+	 * @param pipelineOption runtime parameter passed to pipeline
+	 * @return decrypted value if pipeline parameter is encrypted otherwise returns
+	 *         the raw value as is.
 	 */
-    public static <T extends ValueProvider>String getStringValue(T t){
-        if (t.isAccessible()) {
-            return (String)t.get();
-        }
-        return "";
-    }
-    
-    /**
-     * Checks if pipeline parameter is encrypted. If pipeline parameter is encrypted, returns the decrypted value 
-     * using kms encryption key otherwise returns the raw value as is.
-     * 
-     * @param pipelineParameter runtime parameter passed to pipeline
-     * @param kmsKey the encryption key
-     * @return decrypted value if pipeline parameter is encrypted otherwise returns the raw value as is.
-     */
-    public static String getRawEncrypted(ValueProvider<String> pipelineParameter, String kmsKey) {
-
-    	String pipelineParameterStringValue = getStringValue(pipelineParameter);
-    	
-    	if(isNullOrEmpty(pipelineParameterStringValue)) {
-    		return "";
-    	}
-
-    	String decryptedValue;
-    	if (!isEncrypted(pipelineParameterStringValue)) {
-    		return pipelineParameterStringValue;
-    	}
-    	else {
-    		try {
-				decryptedValue=decryptWithKMS(extractEncryptedValue(getStringValue(pipelineParameter)),kmsKey);
-			} catch (IOException e) {
-				throw new RuntimeException(e);
+	public String decryptIfRequired(ValueProvider pipelineOption) {
+		if (pipelineOption.isAccessible()) {
+			String pipelineOptionValue = (String) pipelineOption.get();
+			if (!Strings.isNullOrEmpty(pipelineOptionValue) && isEncrypted(pipelineOptionValue)) {
+				try {
+					return decryptWithKMS(extractEncryptedValue(pipelineOptionValue));
+				} catch (IOException e) {
+					throw new RuntimeException(e);
+				}
 			}
-   		return decryptedValue;
-    	}
-    }
+			return pipelineOptionValue;
+		}
+		return "";
+	}
 
-    /**
-     * Checks if pipeline parameter is encrypted. If pipeline parameter is encrypted, returns the decrypted value 
-     * using kms encryption key otherwise returns the raw value as is.
-     * 
-     * @param pipelineParameter runtime parameter passed to pipeline
-     * @param kmsKey the encryption key
-     * @return decrypted value if pipeline parameter is encrypted otherwise returns the raw value as is.
-     */
-    public static String getRawEncrypted(String pipelineParameter, String kmsKey) {
-    	
-    	if(isNullOrEmpty(pipelineParameter)) {
-    		return "";
-    	}
-    	
-    	String decryptedValue;
-    	if (!isEncrypted(pipelineParameter)) {
-    		return pipelineParameter;
-    	}
-    	else {
-    		try {
-				decryptedValue=decryptWithKMS(extractEncryptedValue(pipelineParameter),kmsKey);
-			} catch (IOException e) {
-				throw new RuntimeException(e);
-			}
-    		return decryptedValue;
-    	}
-    }
-    
-    /**
-     * Checks whether the argument is encrypted. Encrypted values should be in the following 
-     * format : ${encryptedValue}. The raw values can be passed as is.
-     * 
-     * @param pipelineParameter to check for encryption.
-     * @return true if argument is encrypted
-     */
-    public static boolean isEncrypted(String pipelineParameter) {
-    	String regex="\\$\\{.*?}";
-    	return pipelineParameter.matches(regex);
-    }
+	/**
+	 * Checks if pipeline parameter is encrypted. If pipeline parameter is
+	 * encrypted, returns the decrypted value using kms encryption key otherwise
+	 * returns the raw value as is.
+	 * 
+	 * @param pipelineOption runtime parameter passed to pipeline
+	 * @return decrypted value if pipeline parameter is encrypted otherwise returns
+	 *         the raw value as is.
+	 */
+	public String decryptIfRequired(String pipelineOption) {
+		return decryptIfRequired(StaticValueProvider.of(pipelineOption));
+	}
 
-    /**
-     * Extracts the encrypted value between the template :${}.
-     * 
-     * @param pipelineParameter encrypted value of the following format : ${encryptedValue}
-     * @return extracted value if the argument is encrypted otherwise return the raw value as is.
-     */
-    public static String extractEncryptedValue(String pipelineParameter){
-    	String pattern="\\$\\{(.*?)}";
-    	Matcher m = Pattern.compile(pattern).matcher(pipelineParameter);
-    	while (m.find()) {
-    		return m.group(1);
-    	}
-    	
-    	return "";
-    }
+	/**
+	 * Checks whether the argument is encrypted. Encrypted values should be in the
+	 * following format : ${encryptedValue}. The raw values can be passed as is.
+	 * 
+	 * @param pipelineParameter to check for encryption.
+	 * @return true if argument is encrypted
+	 */
+	public static boolean isEncrypted(String pipelineParameter) {
+		return ENCRYPT_PATTERN.matcher(pipelineParameter).matches();
+	}
 
-    /**
-     * Checks if string argument is null or empty.
-     * 
-     * @param str the string to check
-     * @return true if string is null or empty else returns false.
-     */
-    public static boolean isNullOrEmpty(String str) {
-    	return !(str != null && !str.trim().isEmpty());
-    }
-	
+	/**
+	 * Extracts the encrypted value between the template :${}.
+	 * 
+	 * @param pipelineParameter encrypted value of the following format :
+	 *                          ${encryptedValue}
+	 * @return extracted value if the argument is encrypted otherwise return the raw
+	 *         value as is.
+	 */
+	public static String extractEncryptedValue(String pipelineParameter) {
+		Matcher m = ENCRYPT_PATTERN.matcher(pipelineParameter);
+		while (m.find()) {
+			return m.group(1);
+		}
+		return "";
+	}
+
 }
